fixed dependency issues
